% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calendlyR.R
\docType{package}
\name{calendlyR-package}
\alias{calendlyR}
\alias{calendlyR-package}
\title{calendlyR: Access the calendly API using R.}
\description{
Package details...
}
\details{
Visit the calendly API documentation: \url{https://calendly.stoplight.io/docs/gh/calendly/api-docs/docs/API-Use-Cases.md?srn=gh/calendly/api-docs/docs/API-Use-Cases.md}
}
\author{
\strong{Maintainer}: Bill Ash \email{ash.william.r@gmail.com}

}
